new aggressively normalized version
Added an experimental version of gdual_forward which normalizes much
more aggressively. Updated test_stability to an “unstable” test. That
test fails about 50% of the time, but in about 50% of fail instances,
only the less aggressively normalized variant crashes.

I took one of these cases and hard coded it as an option in
test_stability if you set “sample_counts” to false