Normalized gdual refactor
Refactored gdual and gdualforward to ngdual and ngdualforward which
operate on “normalized gduals”. An ngdual tuple is composed of a log
normalizer and a normalized UTP (normalized to have max(UTP) == 1). I
also added a unit test in ngdual_test.py which tests that all the
ngdual and gdual operations match. It’s an immature unit test (asserts
don’t work due to some instability introduced in converting to/from
logspace, relying instead of manual confirmation). It could use some
more test cases as we go forward, but should be alright for now.

ngdualforward still has a bug somewhere that I am working out.