ngdual pgfs and "invalid val in log" error
Added support for ngdual versions of all pgfs in generatingfunctions.py

Found the source of the frequent “Runtimewarning: invalid value
encountered in log” warning. Z was computed as the max of all terms in
the out, which was sometimes negative, so logZ was undefined. Updated
to that Z is the largest magnitude value (pos or neg), which will mean
some utps are normalized so the largest magnitude value goes to -1. I
don’t think this is a problem.