Change sign to 32 bits so C code can be more easily compatile with Python and R. Reogarnize modules slightly. It seems very hard to support "cross-package" imports, especially when these can become circular.
For example, consider the following example (at least close to something we had):

lsgdual/test_lsgdual --> gdual/gdual --> lsgdual/util

It seems like there is almost no way to get the "import lsgdual.util" statement
correct. In general it should be "import lsgdual.util", but when the script
test_lsgdual.py is run, this is incorrect because the scope is within the lsgdual
package, so it should be "import util".

I suggest we minimize cross-package imports as much as possible. To facilitate
this, I put the cygdual stuff inside lsgdual, which is appropriate because these
are really C versions of the corresponding lsgdual operations.

I also moved util.py to the top level outside of any of the packages.

I did not make this change, but suggest that we put scripts that compare different
implementations outside of the packages as well.