Implementations of Brent-Kung compose algorithm for lsgdual and cygdual. After considerable debugging, the C version works and is fairly fast/stable in initial testing. Also includes a mul operation for different length gduals, and the beginnings of a design for other operations that take different sized operands. However, this is very unlikely to be a major performance consideration, since ~100% of time is still spent in compose.
An interesting performance observation: this is much faster (probably
an order of magnitude) when the offspring PGF is Bernoulli. And we don't
use the "affine" special case. I believe the reason is that the Bernoulli
PGF leads to gduals that are mostly zero, and zeros are handled specially
in the ls_add() operation, which is the absolute inner loop operation,
and is pretty expensive (calls log1p).