Wire up gradient calculations to MLE experiments (for case study only for now). The distribution "objects" all need to implement three additional fields now:
pgf_grad         The gradient version of the pgf (from forward_grad.py)

need_grad        Function to generate mask of which Parameters need
                 gradients from the objective calculation. Should return
                 a list/array w/ same dimensions as corresponding parameters.

backprop         Function to backpropagate parameter gradient to
                 hyperparameter gradient. (Must return a 1d numpy array)

I only populated these for a few objects.

Bad news: the backprop version is ~50% slower than the non-backprop version
for this version of the case study. Quick analysis: there are only three
hyperpameters, so without using backprop we need to call the forward
algorithm 4 times: once to compute the base value and once to compute the
forward difference partial derivative for each of the 3 parameters. When
using backprop, we need to call the compose operation a bunch more times,
up to K times per parameter, so we are multiplying the running time of
forward by more than 4. This calculus could change if there are more
hyperparameters, i.e., less parameter tying in the model, because
that will make the forward difference gradient calculation take much
longer, but the backprop gradient calculation will remain about the same.